Fix Pixel Unit national colors reverting to white after battle
